backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-migrate
sqlalchemy
psycopg2-binary
requests
python-dotenv
pandas
numpy
gunicorn
google-genai
# Segurança / Auth
PyJWT
pyotp

# Add other dependencies as needed
